# Faster frame reading in ServiceProxy