--- a/benchmarks/hyperbahn-worker.js
+++ b/benchmarks/hyperbahn-worker.js
@@ -130,7 +130,7 @@ HyperbahnWorker.prototype.createConfig = function createConfig() {
                 leafPort: self.kafkaPort
             },
             'clients.logtron.logFile': null,
-            // 'clients.logtron.console': true,
+            'clients.logtron.console': true,
             'clients.logtron.sentry': {
                 id: 'http://bs:bs@localhost:' + self.sentryPort
             },